% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_prov.R
\name{div_prov}
\alias{div_prov}
\title{Amount instance of disease by province}
\usage{
div_prov(x)
}
\arguments{
\item{x}{vector of length three (illness or disease, province, year)}
}
\value{
value with total cases by province and year
}
\description{
Amount instance of disease by province
}
\examples{
div_prov("Dengue","ORELLANA", 2018)
}
