% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_week.R
\name{div_week}
\alias{div_week}
\title{Principal measures of disease cases by epidemiological week}
\usage{
div_week(SEe)
}
\arguments{
\item{SEe}{Epidemiological week}
}
\value{
Total cases in this epidemiological week
}
\description{
Principal measures of disease cases by epidemiological week
}
\examples{
For this is necessary a data frame with an specifically disease and year.
div_week(54)
}
