% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_year.R
\name{div_year}
\alias{div_year}
\title{Amount instance of disease by year}
\usage{
div_year(x)
}
\arguments{
\item{x}{vector of length two (illness or disease, year)}
}
\value{
data frame with total cases by year
}
\description{
Amount instance of disease by year
}
\examples{
div_year("Malaria",2016)
}
