% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diss.R
\name{plot_diss}
\alias{plot_diss}
\title{Graphic representation of a disease per year}
\usage{
plot_diss(x)
}
\arguments{
\item{x}{value of disease ("Malaria")}
}
\value{
bar graphic of frecuenciy of a disease per year
}
\description{
Graphic representation of a disease per year
}
\examples{
plot_diss("Malaria")
}
